+++ pkg/domain/accounts/check_accounts_test.go
@@ -4,14 +4,14 @@ import (
 	"testing"
 
 	"exemplo.com/pkg/domain/entities"
-	"exemplo.com/pkg/store"
+	"exemplo.com/pkg/store/accounts"
 )
 
 func TestAccountUseCase_CheckAccounts(t *testing.T) {
 
 	t.Run("should return nil when accounts have already been created", func(t *testing.T) {
 
-		storage := store.NewAccountStorage()
+		storage := accounts.NewAccountStorage()
 		AccountUseCase := NewAccountUseCase(storage)
 		account, err := entities.NewAccount("John Doe", "11111111030", "123", 10)
 		if err != nil {
@@ -30,7 +30,7 @@ func TestAccountUseCase_CheckAccounts(t *testing.T) {
 
 	t.Run("should return an error message when id isn't found", func(t *testing.T) {
 
-		storage := store.NewAccountStorage()
+		storage := accounts.NewAccountStorage()
 		AccountUseCase := NewAccountUseCase(storage)
 		account, err := entities.NewAccount("John Doe", "11111111030", "123", 10)
 		if err != nil {
