+++ pkg/domain/accounts/get_accounts_test.go
@@ -4,14 +4,14 @@ import (
 	"testing"
 
 	"exemplo.com/pkg/domain/entities"
-	"exemplo.com/pkg/store"
+	"exemplo.com/pkg/store/accounts"
 )
 
 func TestAccountUseCase_GetAccounts(t *testing.T) {
 
 	t.Run("should return a full list of accounts", func(t *testing.T) {
 
-		storage := store.NewAccountStorage()
+		storage := accounts.NewAccountStorage()
 		AccountUseCase := NewAccountUseCase(storage)
 		account, err := entities.NewAccount("John Doe", "11111111030", "123", 10)
 		if err != nil {
@@ -29,7 +29,7 @@ func TestAccountUseCase_GetAccounts(t *testing.T) {
 
 	t.Run("should return an empty list", func(t *testing.T) {
 
-		storage := store.NewAccountStorage()
+		storage := accounts.NewAccountStorage()
 		AccountUseCase := NewAccountUseCase(storage)
 
 		getAccounts := AccountUseCase.GetAccounts()
