+++ pkg/store/accounts/join_accounts.go
@@ -1,4 +1,4 @@
-package store
+package accounts
 
 import "exemplo.com/pkg/domain/entities"
