+++ pkg/store/accounts/new_account_storage.go
@@ -1,4 +1,4 @@
-package store
+package accounts
 
 import (
 	"errors"
